pyproject.toml
[tool.poetry]
name = "employees"
version = "0.1.0"
description = "A list of employees"
authors = ["Vova <bkj@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
python-dotenv = "^0.20.0"
dj-database-url = "^1.0.0"
pip = "^22.2.2"
psycopg2-binary = "^2.9.3"
django-seed = "^0.3.1"
django-extensions = "^3.2.0"
django-bootstrap4 = "^22.2"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.13.1"
Markdown = "^3.4.1"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
